Fix up bug in getting total sum abiut paid with frontend redirection from login to verification of otp

+++ src/main/java/com/boyboys/dues_payment_system/reports/domain/OverallSummaryResponse.java
@@ -1,13 +1,14 @@
 package com.boyboys.dues_payment_system.reports.domain;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OverallSummaryResponse(
         long totalStudents,
         long totalPaid,
         long totalUnpaid,
-        long totalAmountCollectedInCedis,
+        BigDecimal totalAmountCollectedInCedis,
         List<ProgrammeSummary> programmeSummaries,
         List<LevelSummary> levelSummaries
 ) {}
